+++ workshop_4/chatbot/.gitignore
@@ -1,2 +1,1 @@
-package
 *.zip
